requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
requests>=2.28.0
pillow>=9.0.0
matplotlib>=3.8.0